v2.0.0 | Переход на FastAPI
