+++ requirements.txt
@@ -1,2 +1,4 @@
 selenium
-flask
+fastapi
+pydantic
+uvicorn
